Move logic to main box
